+++ ui-new/app/login/page.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useState } from "react";
 import useLogin from "@/hooks/useLogin";
 import { useRouter } from "next/navigation";
-import { Button } from "@/components/extras/buttons";
+import { Button } from "@/components/ui/buttons";
 import Link from "next/link";
 
 function Login() {
